requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph>=0.0.65
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
slack-sdk>=3.26.1
jira>=3.5.2
python-pptx>=0.6.23
aiofiles>=0.24.0
httpx>=0.25.2
aiohttp>=3.8.0
Pillow>=10.1.0
mcp>=1.0.0
mcp-client>=1.0.0